requirements.txt
accessible-pygments==0.0.5
alabaster==0.7.16
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
blimpy
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
docutils==0.21.2
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
imagesize==1.4.1
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
iqrm
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-cache==1.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
livereload==2.6.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
myst-nb==1.1.0
myst-parser==3.0.1
nbclient==0.10.0
nbformat==5.10.4
nest-asyncio==1.6.0
numba
numpydoc==1.7.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydata-sphinx-theme==0.15.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
scipy
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.3.7
sphinx-autobuild==2021.3.14
sphinx-book-theme==1.1.3
sphinx-thebe==0.3.1
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mermaid==0.9.2
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
tabulate==0.9.0
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
zipp==3.19.2
